db/requirements.txt
Django==1.8.6
Jinja2==2.8
MarkupSafe==0.23
Pygments==2.0.2
argparse==1.2.1
binaryornot==0.4.0
chardet==2.3.0
click==5.1
cookiecutter==1.3.0
django-rest-auth==0.6.0
djangorestframework==3.3.1
future==0.15.2
psycopg2==2.6.1
ruamel.base==1.0.0
ruamel.ordereddict==0.4.9
ruamel.yaml==0.10.12
six==1.10.0
whichcraft==0.1.1
wsgiref==0.1.2
